+++ src/core/datatypes/primitive/Boolean.h
@@ -30,30 +30,30 @@ class VectorString;
 class Boolean : public RbObject {
 
     public:
-        friend class            VectorBoolean;                                                  //!< Give VectorBoolean direct access to value
+        friend class            VectorBoolean;                                                          //!< Give VectorBoolean direct access to value
 
-                                Boolean(void);                                                  //!< Default constructor
-                                Boolean(const bool v);                                          //!< Construct from bool
+                                Boolean(void);                                                          //!< Default constructor
+                                Boolean(const bool v);                                                  //!< Construct from bool
 
         // Operators
-                                operator bool(void) const { return value; }                     //!< Type conversion to bool
+                                operator bool(void) const { return value; }                             //!< Type conversion to bool
 
         // Basic utility functions
-        Boolean*                clone(void) const;                                              //!< Clone object
-        RbObject*               convertTo(const std::string& type, int dim) const;              //!< Convert to type and dim
-        const VectorString&     getClass(void) const;                                           //!< Get class vector
-        bool                    isConvertibleTo(const std::string& type, int dim, bool once) const;     //!< Is convertible to type and dim?
-        void                    printValue(std::ostream& o) const;                              //!< Print value (for user)
-        std::string             richInfo(void) const;                                           //!< Complete info about object
+        Boolean*                clone(void) const;                                                      //!< Clone object
+        RbObject*               convertTo(const std::string& type, size_t dim) const;                   //!< Convert to type and dim
+        const VectorString&     getClass(void) const;                                                   //!< Get class vector
+        bool                    isConvertibleTo(const std::string& type, size_t dim, bool once) const;  //!< Is convertible to type and dim?
+        void                    printValue(std::ostream& o) const;                                      //!< Print value (for user)
+        std::string             richInfo(void) const;                                                   //!< Complete info about object
 
         // Getters and setters
-        void                    setValue(bool x) { value = x; }                                 //!< Set value
-        bool                    getValue(void) const { return value; }                          //!< Get value
+        void                    setValue(bool x) { value = x; }                                         //!< Set value
+        bool                    getValue(void) const { return value; }                                  //!< Get value
 
     private:
-        bool&                   getValueRef(void) { return value; }                             //!< Get value reference
+        bool&                   getValueRef(void) { return value; }                                     //!< Get value reference
 
-        bool                    value;                                                          //!< Value member
+        bool                    value;                                                                  //!< Value member
 };
 
 #endif
